Handle compiler warnings.
Mostly unused variables, some more esoteric stuff as well.

Warnings left in drivers:
  - hw_irman: bogus warning for variable text set-but-not-used.
  - hw_commandir: countless, too many to fix.
  - below in hw_audio_alsa. Potentially dangerous and should be
    fixed, but by someone who knows this better.

hw_audio_alsa.c: In function ‘alsa_sig_io’:
hw_audio_alsa.c:388:11: warning: dereferencing type-punned
    pointer will break strict-aliasing rules [-Wstrict-aliasing]
           bytes_per_sample * alsa_hw.channel]) >> 8);

Other warnings: man2html: odd, really odd #ifdefs creates warning.

Fix compiler warnings

+++ daemons/hw_audio.c
@@ -99,7 +99,7 @@ static void addCode(lirc_t data)
 ** that could mess up the system like calling malloc() or free().
 */
 
-static int recordCallback(const void *inputBuffer, void *outputBuffer, __u32 framesPerBuffer,
+static int recordCallback(const void *inputBuffer, void *outputBuffer, unsigned long framesPerBuffer,
 			  const PaStreamCallbackTimeInfo * outTime, PaStreamCallbackFlags status, void *userData)
 {
 	paTestData *data = (paTestData *) userData;
